+++ flows.yaml
@@ -11215,6 +11215,28 @@ integrations:
                     method: GET
                     path: /emails
                 version: 1.1.0
+            events:
+                description: |
+                    Sync calendar events on the primary calendar going back one month and
+                    save the entire object as specified by the Outlook Calendar API
+                runs: every 5 minutes
+                output: OutlookCalendarEvent
+                sync_type: incremental
+                endpoint:
+                    method: GET
+                    path: /outlook-calendar/events
+                scopes: Calendars.Read
+            calendars:
+                description: |
+                    Sync the calendars list of the user
+                track_deletes: true
+                runs: every hour
+                output: OutlookCalendar
+                sync_type: full
+                endpoint:
+                    method: GET
+                    path: /outlook-calendar/calendars
+                scopes: Calendars.Read
         actions:
             fetch-attachment:
                 input: DocumentInput
@@ -11246,31 +11268,6 @@ integrations:
             DocumentInput:
                 threadId: string
                 attachmentId: string
-    outlook-calendar:
-        syncs:
-            events:
-                description: |
-                    Sync calendar events on the primary calendar going back one month and
-                    save the entire object as specified by the Outlook Calendar API
-                runs: every 5 minutes
-                output: OutlookCalendarEvent
-                sync_type: incremental
-                endpoint:
-                    method: GET
-                    path: /outlook-calendar/events
-                scopes: Calendars.Read
-            calendars:
-                description: |
-                    Sync the calendars list of the user
-                track_deletes: true
-                runs: every hour
-                output: OutlookCalendar
-                sync_type: full
-                endpoint:
-                    method: GET
-                    path: /outlook-calendar/calendars
-                scopes: Calendars.Read
-        models:
             OutlookCalendar:
                 id: string
                 allowedOnlineMeetingProviders: string[]
@@ -11287,81 +11284,80 @@ integrations:
                 isRemovable: boolean
                 isTallyingResponses: boolean
                 name: string
-                ownerEmailAddress: string
-                ownerName: string
+                owner: EmailAddress
             OutlookCalendarEvent:
                 id: string
                 allowNewTimeProposals?: boolean
                 attendees: Attendee[]
-                body:
-                    content: string
-                    contentType: text | html
+                body?:
+                    content?: string
+                    contentType?: text | html
                 bodyPreview: string
-                categories: string[]
-                changeKey: string
-                createdDateTime: string
+                categories?: string[]
+                changeKey?: string
+                createdDateTime?: string
                 end:
                     dateTime: string
                     timeZone: string
-                hasAttachments: boolean
-                hideAttendees: boolean
-                iCalUId: string
+                hasAttachments?: boolean
+                hideAttendees?: boolean
+                iCalUId?: string
                 importance: low | normal | high
                 isAllDay: boolean
                 isCancelled: boolean
-                isDraft: boolean
+                isDraft?: boolean
                 isOnlineMeeting?: boolean
                 isOrganizer: boolean
-                isReminderOn: boolean
-                lastModifiedDateTime: string
+                isReminderOn?: boolean
+                lastModifiedDateTime?: string
                 location: Location
-                locations: Location[]
+                locations?: Location[]
                 onlineMeeting: OnlineMeetingInfo | null
                 onlineMeetingProvider?: string
                 onlineMeetingUrl?: string
                 organizer:
                     emailAddress: EmailAddress
-                originalEndTimeZone: string
-                originalStart: string
-                originalStartTimeZone: string
-                recurrence: PatternedRecurrence
-                reminderMinutesBeforeStart: number
+                originalEndTimeZone?: string
+                originalStart?: string
+                originalStartTimeZone?: string
+                recurrence?: PatternedRecurrence | null
+                reminderMinutesBeforeStart?: number
                 responseRequested: boolean
                 responseStatus:
-                    response: none | accepted | declined | tentative
-                    sentDateTime: string
+                    response: string
+                    time: string
                 sensitivity: normal | personal | private | confidential
-                seriesMasterId: string
-                showAs: free | tentative | busy | oof | workingElsewhere | unknown
+                seriesMasterId?: string
+                showAs?: free | tentative | busy | oof | workingElsewhere | unknown
                 start:
                     dateTime: string
                     timeZone: string
                 subject: string
                 transactionId?: string
-                type: singleInstance | occurrence | exception | seriesMaster
+                type?: singleInstance | occurrence | exception | seriesMaster
                 webLink: string
             Location:
-                address: string
-                city: string
-                countryOrRegion: string
-                postalCode: string
-                state: string
-                street: string
-                coordinates: null
-                accuracy: number
-                altitude: number
-                altitudeAccuracy: number
-                latitude: number
-                longitude: number
-                displayName: string
-                locationEmailAddress: string
-                locationUri: string
-                locationType: >-
+                address?:
+                    city?: string
+                    countryOrRegion?: string
+                    postalCode?: string
+                    state?: string
+                    street?: string
+                coordinates?:
+                    accuracy?: number
+                    altitude?: number
+                    altitudeAccuracy?: number
+                    latitude?: number
+                    longitude?: number
+                displayName?: string
+                locationEmailAddress?: string
+                locationUri?: string
+                locationType?: >-
                     default | conferenceRoom | homeAddress | businessAddress |
                     geoCoordinates | streetAddress | hotel | restaurant | localBusiness |
                     postalAddress
-                uniqueId: string
-                uniqueIdType: string
+                uniqueId?: string
+                uniqueIdType?: string
             OnlineMeetingInfo:
                 conferenceId: string
                 joinUrl: string
