+++ integrations/outlook/mocks/calendars/OutlookCalendar/batchDelete.json
@@ -1,0 +1,1 @@
+[]
