+++ integrations/outlook/mocks/calendars/OutlookCalendar/batchSave.json
@@ -1,0 +1,21 @@
+[
+    {
+        "id": "AAMkADFmM2Y5OTNlLWI1YmEtNDY3Ny1iMzE5LTNkYzkxMWE3OTQ5NwBGAAAAAADe1HLlVCjpQ7N9oqf6MQi_BwBnb1obBfp9TL8KB3yxQVrvAAAAAAEGAABnb1obBfp9TL8KB3yxQVrvAAFx6xuKAAA=",
+        "name": "Calendar",
+        "color": "auto",
+        "hexColor": "",
+        "isDefaultCalendar": true,
+        "changeKey": "Z29aGwX6fUy/Cgd8sUFa7wABcepvgg==",
+        "canShare": true,
+        "canViewPrivateItems": true,
+        "canEdit": true,
+        "allowedOnlineMeetingProviders": ["teamsForBusiness"],
+        "defaultOnlineMeetingProvider": "teamsForBusiness",
+        "isTallyingResponses": true,
+        "isRemovable": false,
+        "owner": {
+            "name": "Khaliq Gant",
+            "address": "KhaliqGant@nangodev.onmicrosoft.com"
+        }
+    }
+]
