+++ integrations/outlook/mocks/events/OutlookCalendarEvent/batchDelete.json
@@ -1,0 +1,1 @@
+[]
